requirements.txt
asgiref==3.8.1
awsebcli==3.23.2
blessed==1.21.0
botocore==1.38.4
cement==2.10.14
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
idna==3.10
jmespath==1.0.1
packaging==24.2
pathspec==0.12.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
setuptools==80.0.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.13
langchain-community
pandas
openai
python-dotenv
numpy
requests
tiktoken
pypdf2
ortools

